feat(系统设置): 模型设置高级设置
