+++ frontend/src/assets/style/arco-reset.less
@@ -233,6 +233,12 @@
   .btn-base-danger-active();
   .btn-base-danger-disabled();
 }
+.arco-btn-status-warning {
+  background-color: rgb(var(--warning-6)) !important;
+  .btn-base-warning-hover();
+  .btn-base-warning-active();
+  .btn-base-warning-disabled();
+}
 .arco-btn-secondary:not(.arco-btn-disabled) {
   color: var(--clolor-text-1) !important;
   background-color: var(--color-text-n8) !important;
