+++ frontend/src/components/business/ms-batch-form/types.ts
@@ -27,5 +27,7 @@ export interface FormItemModel {
   hasRedStar?: boolean; // 是否有红星
   tooltip?: string;
   disabled?: boolean;
+  maxKey?: string;
+  minKey?: string;
   [key: string]: any;
 }
